Translated Constant.Int type to long long data type in C and generates the loop condition by merging the range and forall WyIL bytecode.

+++ tests/code/valid/BoolList_Valid_2.h
@@ -5,7 +5,7 @@
 #define false 0
 #define true 1
 #define NULL -1
-int* play(int* _0, int _0_size);
-int isAlive(int _0, int* _1, int _1_size);
+long long* play(long long* _0, long long _0_size);
+long long isAlive(long long _0, long long* _1, long long _1_size);
 void main(int argc, char** argv);
